change to sync multiple calendar

+++ google-calendar-register/src/main/java/com/tq/googlecalendar/lambda/handler/GoogleDisconnectCalendarHandler.java
@@ -1,8 +1,11 @@
 package com.tq.googlecalendar.lambda.handler;
 
+import java.util.List;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.tq.common.lambda.dynamodb.model.GCModifiedChannel;
 import com.tq.common.lambda.dynamodb.model.GoogleCalendarSbmSync;
 import com.tq.common.lambda.dynamodb.model.GoogleRenewChannelInfo;
 import com.tq.common.lambda.dynamodb.service.GoogleCalRenewService;
@@ -13,8 +16,8 @@ import com.tq.googlecalendar.exception.GoogleApiSDKException;
 import com.tq.googlecalendar.impl.GoogleCalendarApiServiceBuilder;
 import com.tq.googlecalendar.impl.TokenGoogleCalendarImpl;
 import com.tq.googlecalendar.lambda.exception.TrueQuitRegisterException;
-import com.tq.googlecalendar.lambda.model.GoogleConnectStatusResponse;
 import com.tq.googlecalendar.lambda.model.GoogleRegisterReq;
+import com.tq.googlecalendar.lambda.resp.GoogleConnectStatusResponse;
 import com.tq.googlecalendar.req.StopWatchEventReq;
 import com.tq.googlecalendar.req.TokenReq;
 import com.tq.googlecalendar.resp.TokenResp;
@@ -49,37 +52,45 @@ public class GoogleDisconnectCalendarHandler implements Handler {
 			throws GoogleApiSDKException, TrueQuitRegisterException {
 		GoogleConnectStatusResponse response = new GoogleConnectStatusResponse();
 		String sbmEmail = req.getParams().getEmail();
-		GoogleCalendarSbmSync googleCalendarSbmSync = googleCalendarService.query(sbmEmail);
-		if (googleCalendarSbmSync != null) {
-			TokenReq tokenReq = new TokenReq(eVariables.getGoogleClientId(), eVariables.getGoogleClientSecrets(),
-					googleCalendarSbmSync.getRefreshToken());
+		List<GoogleCalendarSbmSync> googleCalendarSbmSync = googleCalendarService.queryEmail(sbmEmail);
+		if (!googleCalendarSbmSync.isEmpty()) {
+			for (GoogleCalendarSbmSync googleSbm : googleCalendarSbmSync) {
 
-			TokenResp tokenResp = tokenCalendarService.getToken(tokenReq);
-			StopWatchEventReq stopEventReq = new StopWatchEventReq(googleCalendarSbmSync.getSbmId(),
-					googleCalendarSbmSync.getGcWatchResourceId());
-			GoogleCalendarApiService googleApiService = apiServiceBuilder.build(tokenResp.getAccess_token());
-			boolean flag = false;
-			// work-around: can' stop channel in just one request. So as to make sure the
-			// channel is stopped successfully, we try to stop maximum of 3 times .
-			boolean isChecked = GoogleCalendarUtil.stopWatchChannel(googleApiService, stopEventReq, flag);
-			if (isChecked) {
-				googleCalendarService.delete(googleCalendarSbmSync);
-				m_log.info("Delete record in table GoogleCalendarSbmSync successfully");
-				GoogleRenewChannelInfo renewChannel = googleCalRenewService.query(googleCalendarSbmSync.getSbmId());
-				if (renewChannel != null) {
+				TokenReq tokenReq = new TokenReq(eVariables.getGoogleClientId(), eVariables.getGoogleClientSecrets(),
+						googleSbm.getRefreshToken());
+
+				TokenResp tokenResp = tokenCalendarService.getToken(tokenReq);
+				StopWatchEventReq stopEventReq = new StopWatchEventReq(googleSbm.getChannelId(),
+						googleSbm.getGcWatchResourceId());
+				GoogleCalendarApiService googleApiService = apiServiceBuilder.build(tokenResp.getAccess_token());
+				GoogleRenewChannelInfo renewChannel = googleCalRenewService.queryChannelId(googleSbm.getGcWatchResourceId());
+				boolean flag = false;
+				// work-around: can' stop channel in just one request. So as to make sure the
+				// channel is stopped successfully, we try to stop maximum of 3 times .
+				boolean isChecked = GoogleCalendarUtil.stopWatchChannel(googleApiService, stopEventReq, flag);
+				if (isChecked && renewChannel != null) {
+					googleCalendarService.delete(googleSbm);
+					m_log.info("Delete record in table GoogleSbmSync successfully");
 					googleCalRenewService.deleteItem(renewChannel);
-					m_log.info("Delete record in table GoogleRenewChannelInfo successfully");
-					m_log.info("Channel Id" + googleCalendarSbmSync.getSbmId());
-					calendarModifiedChannelService.deleteDynamoItem(googleCalendarSbmSync.getSbmId());
-					m_log.info("Delete record in table GoogleModified successfully");
-
 				}
-
-			} else {
+				else {
+					throw new GoogleApiSDKException("Internal error");
+				}
+			}
+			List<GCModifiedChannel> modifiedItems = calendarModifiedChannelService.queryEmail(sbmEmail);
+			if (!modifiedItems.isEmpty()) {
+				calendarModifiedChannelService.deleteModifiedItem(modifiedItems);
+			} else
+				
+			{
 				throw new GoogleApiSDKException("Internal error");
 			}
 
-		} else {
+		}
+
+		else
+
+		{
 			throw new TrueQuitRegisterException("The email " + sbmEmail + " is not connected yet ");
 		}
 		response.setSucceeded(true);
